Extract rest routes into a new controller class.
